style(core): Fix remaining lint errors (#86)
* refactor(core): Move validate package to internal

As we deferred any anything except our proof-of-concept validation,
this is definitely something we should keep as an implementation
detail until we've thought more about it. (We potentially should
even delete it as it's a bit strange that we're only validating
one field, but at least let's make it internal.)

Also fix the lint failure, which doesn't like repetition in
package/function names (validate.ValidateHalConfig).

Finally change some names to lowercase so they aren't exported
as they aren't used outside the package.

* style(core): Add documentation to fix lint errors

Add documentation to exported members, fixing the last lint
errors in kleat.

* style(core): Fix inspections

I ran the IntelliJ inspections, and it found three issues, which
I fixed since I'm on a roll with minor cleanup. They are:
* Explicitly assign the result of fmt.Printf to _ to indicate that
we are ignoring the error (as there's not much point handling it
right before os.Exit anyway). Also use Printf instead of Fprintf
as we aren't formatting
* Rename 'diff' variable which collides with package name
* Change unused 'h' to _ in function signature (though obviously this
will need to be updated when the TODO is handled).

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>