feat(proto): Compile protoc-gen-doc instead of using release
We'll need to temporarily use a fork of protoc-gen-doc so that
we can include optional fields in proto3. As I don't want to actually
make official releases of the fork, we'll want to build in the
container from a specific tag/commit.

This is what we used to do, before I changed it to pull the release
(which is faster). This will be a bit slower the first time you
run 'make proto' but should be temporary. Once the changes are in
a released version upstream, we can go back to pulling an official
release.

If for some reason these changes don't end up upstream and it becomes
an issue, I can make a release in my fork, but want to avoid that
assuming this is temporary.