feat(proto): Pull version of protoc-gen-doc allowing proto3 optional
This commit pulls in a version of protoc-gen-doc that allows proto3
optional protos. I've created a tag on my fork v1.3.2-proto3optional.1
corresponding to the changes I've submitted upstream. (The idea being
that any other changes would be come v1.3.2-proto3optional.2, and
if I rebase off of v1.3.3 it would become v1.3.3-proto3optional.1
but hopefully this will get merged quickly upstream so we won't need
to update this much).