feat(release): add docs, windows release target, and fix ldflags for version command (#92)
* feat(docs): add Installation and Usage sections to README

* fix(release): qualify ldflags packages as full import paths

When ldflags are intended to be injected in non-main packages, the package name must be fully qualified.

* feat(release): include windows as $GOOS target

Kleat will now be released for the matrix of $GOOS (darwin, linux, windows) and $GOARCH (386, amd64), which are all valid combinations per https://golang.org/doc/install/source#environment. Added documentation to .goreleaser.yml in case users need additional combinations in the future.