test(core): Add an integration test on config generation
We have two tests currently that test config generation:
* A unit test that ensures that the contents of config.Services
is the same as individually calling each of the service config
generators.
* An integration test that ensures the serialized YAML for each
service matches the expected result in testdata.

Let's add a test intermediate between these that does the following:
* Takes test halconfig
* Generates the service configs as protos
* Ensures these protos match those obtained by deserialzing the
data in the test directory.

Note that we're not actually testing serialization here, just using
YAML as a convenient format for describing the expected output of
the integration test. (This is why we read deck-env.yml and deck.yml
as YAML files even though the actual output of kleat here is not
YAML.)