test(core): Update config tests to compare strings
Currently the tests on config generation run both the expected and
actual config through a YAML serializer. This means that errors
in the serialization (such as dropping a field) would not be
caught. This is particularly important for fields that we want
to include even when set to their zero value; these tests will
not fail if that logic is implemented incorrectly.

Also these tests re-implement the serialization logic, instead of
calling the new GenerateConfigFiles.

Let's fix the test to call GenerateConfigFiles, then compare the
result directly against the expected result. It appears that the
YAML is stable so the test shouldn't be flaky. (This is in contrast
to protojson which deliberately has unstable output.)

Because settings.js uses the output of protojson, and deck.env
depends on the (non-deterministic) iteration order over a protobuf
message fields, we'll omit those from the test.