refactor(proto): Wrap all boolean fields
We've decided that all boolean fields should have presence
information.

Without this, booleans will not be included in the serialized
config if they are 'false'. While this is fine for fields that
default to false (for now), it breaks for any field that
defaults to true. It also would cause issues if any field
were changed in the future to default to true.

One of the advantages of kleat is that it only includes fields
that are explicitly set by the user in its config, so it is
easier to change field defaults, as only users who have explicitly
set a field will be overriding the default.

But if we store a boolean without presence, it's impossible to
ever in the future default it to 'true'. It also might be confusing
for users that some fields in their hal config are not in their
microservice configs.