fix(core): Remove TODO and add documentation
Remove a TODO, and also fix the code which should pass through the
value when non-nil (not only when true).

Also document how to choose a field type. This also re-wraps the file.
I didn't otherwise change anythign above the added section, so didn't
break this into commits as it should be easy to see what was
re-wrapped vs what was added.