refactor(proto): Wrap all boolean fields (#96)
* refactor(proto): Wrap all boolean fields

We've decided that all boolean fields should have presence
information.

Without this, booleans will not be included in the serialized
config if they are 'false'. While this is fine for fields that
default to false (for now), it breaks for any field that
defaults to true. It also would cause issues if any field
were changed in the future to default to true.

One of the advantages of kleat is that it only includes fields
that are explicitly set by the user in its config, so it is
easier to change field defaults, as only users who have explicitly
set a field will be overriding the default.

But if we store a boolean without presence, it's impossible to
ever in the future default it to 'true'. It also might be confusing
for users that some fields in their hal config are not in their
microservice configs.

* fix(core): Remove TODO and add documentation

Remove a TODO, and also fix the code which should pass through the
value when non-nil (not only when true).

Also document how to choose a field type. This also re-wraps the file.
I didn't otherwise change anythign above the added section, so didn't
break this into commits as it should be easy to see what was
re-wrapped vs what was added.

* test(orca): Fix broken test

I fixed the code so explicitly set false is passed through for
pipeline templates, but didn't fix the test.