feat(proto): Add unspecified to enum fields
We want to distinguish unspecified enums from enums set to their
default value. Let's do this by making the default value in each
enum be a sentinel UNSPECIFIED value. This means that when the
field is not specified, it will be deseriazlied as UNSPECIFIED,
and will be omitted from re-serialization. Any valid value will
be included in serialization.