feat(ci): Remove default exclusions from golang-ci
It turns out some things (including commenting exported functions)
are turned off by default on golang-ci. (And I've already forgotten
to comment three exported functions in the pas few days.) Let's turn
off the default exclusions, as we should be starting from a clean
slate.

If it turns out there is something it flags that we find annoying and
don't want to fix, we can figure out how to turn that off individually.

I'm making this commit separately as I want to ensure that it catches
the errors I think it should, then will fix in a follow-up commit.