style(core): Add some missing comments, and fix package
Add comments to the exported functions that are missing comments.
Also, I realize that I used the older version of the proto library
when writing the helper wrapper functions. This worked because
wrappers in the old library just aliases BoolValue to the new
library, but let's remove the indirection (and direct dependence
on the new library).