test(echo): Refactor echo tests to handle stats changes
The stats block has a property that no other field in the hal config
has; even when nothing is specified in the input, we still generate
some output.

This is because one of the blocks in the stats config is the deployment
method, which kleat needs to fill in with details about itself. While
this is a departure from the usual situation where empty input leads
to empty output, it feels like this is the one case where an exception
is warranted given that this output needs to be populated by kleat rather
than by the user.

To make the diff on that change smaller, refactor the tests now to
include this default value for the Stats field (currently nil).