refactor(stats): Use a different message for input/output stats
While currently the only configurable stats field is whether it
is enabled, we'll soon be adding more fields. Some fields will only
be in the output (ie, that Spinnaker is being deployed by kleat and
the kleat version) and should not be in the input. To allow us to
do this, use a different message for the block in the hal config and
in the echo config.

Another approach would be to use the same message, but have the fields
be ignored in the input. I didn't do that as I don't want those documented
as they might add confusion, but am definitely open to further thoughts
on this.