fix(protoc): Fix file permissions in kleat-protobuilder
This commit updates the intermediate containers used to fetch protoc
and its dependencies to no longer run as root.

The first change is to update the 'curl' container to run as a
non-root user 'curl' to set the work directory to this user's
home directory.

After downloading and unzipping the desired files, we run chmod
to set the desired permissions, which in many cases are not set
to what we want in the tar/zip archive. We want 755 for directories
and executable files and 644 for non-executable files, so that only
the owner can modify but anyone can use the files when running
protoc.

When copying the files to the kleat-protobuilder container,
explicitly have them owned by root (leaving the permissions as what
we set them to after downloading). This matches the way things are
normally installed on linux, where root owns the files and is the
only user with permission to write, but any user can read/execute
the files.

In the case of genproto, we'll explicitly follow up with a chmod
to ensure it has mode 755 as we can't necessarily control what
the user has on their system.

As of this commit, we've fixed the permissions so that it is now
possible to run the kleat-protobuilder as a non-root user by
passing the -u flag to docker, but we're not actually doing that
yet, which will follow in an upcoming commit.