fix(protoc): Run protoc container as invoking user
This commit adds the -u flag to the docker command running the
protocol buffer compilation, so that the container runs with the
same uid/gid as the invoking user. This will cause any output files
to have this uid/gid (instead of being owned by root as before).

I've also created a user 'protoc' in the container and set the
user to that user. This is really just a fallback so we're falling
back to a non-root user; this will always be overridden by the
--user flag that we've added to the docker run command.