feat(dependencies): Add dependabot config (#123)
Let's use dependabot to keep our go dependencies up to date, as
well as the base images in our Dockerfile for compiling protos.

This will *not* upgrade the versions of the following in our build
container, which we'll still need to do manually:
* protoc
* protoc-gen-go
* protoc-gen-doc

But it still seems worthwile to have some of these dependencies
auto-bumped for us.