chore(deps): Update some dependencies (#139)
* chore(deps): Update dependencies and run go mod tidy

The dependencies should be updated by dependabot, but I think that
it doesn't add the indirect dependency so that appeared when I ran
go get -u -t ./...

Also I think it doesn't run go mod tidy, so running that manually.

* chore(deps): Upgrade protoc

Upgrade to protoc 3.13.0; this does not change any of the
generated code, only the comment string.