chore(deps): Upgrade golangci-lint
Also, the documentation recommends running it separately from
other jobs so that they run in parallel.  The main reason I'm
doing that here is because there are a lot of warnings in the
logs that I think are because it is expecting to start with
a clean working environment, which is not the case when it
runs after our other checks.