WIP: Example of how to generate kustomization
This commit is a quick proof of concept of how we might use the
kustomize library to generate a kustomization file for users.

This might be useful if users find it cumbersome to maunually
pin the version of each microservice themselves; we could instead
allow them to specify a single version in the halconfig and
output a kustomization for them that pins each microservice to
that version.